[201811][dhcp] enable building dhcpmon and dhcp6relay debug symbols (#9959)
* dhcp6relay: Save the dbgsym package into the target folder (#9013)

This makes it possible to install the debug symbols if needed. Also install
the package into the debug version of sonic-dhcp-relay container.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Package debugging and hardening for dhcpmon and dhcp6relay (#9862)

Enable dbgsym package for dhcpmon.

Allow CFLAGS and LDFLAGS from environment variables to be used
in the dhcp6relay build. This makes sure that the -O2 flag from
dpkg-buildflags gets used.

Finally, enable all hardening flags in dpkg-buildflags for
dhcp6relay and dhcpmon. The change from the default set of flags is that
during linking, immediate binding of symbols is done instead of lazy
binding.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* [201811][dhcp] update debian build rules

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

Co-authored-by: Saikrishna Arcot <sarcot@microsoft.com>