Update template pull_request_template.md (#9888)
Updating template description to include link to configuration schema for YANG module changes.

#### Why I did it
Updating template description to include link to configuration schema for YANG module changes. When reviewing yang models it becomes difficult to know the entire schema for the table for which yang is defined. Besides providing a clear picture, it will also help to document the config_db schema which is not upto date.

#### How I did it
Updating template file