Revert "[build]: Use zstd compression for base filesystem squashfs" (#10030)
When upgrading from an older image (with an older kernel that doesn't
support zstd) to 202012, the squashfs filesystem from the 202012 image
can't be mounted on the older image, because zstd was introduced in
4.14, and older images will likely be running 4.9.

Therefore, disable zstd compression in the squashfs image for 202012
image. When upgrading from 202012 to newer images, since 202012 is
running on 4.19, it can read and mount zstd-compressed squashfs images.

# Why this was done

The kernel in 201811 image (4.9 kernel) doesn't support zstd. In some cases,
it may be needed to mount the squashfs image from the new image while
the current (old) image is booted. If the squashfs image is compressed with
zstd, then it cannot be mounted on a 4.9 kernel, because zstd was introduced
in 4.14.

This reverts commit 35e88e5f6a743422496d79540e46e4cb464186be.